refactor name-related methods to intermediate class
